+++ R/socioecon_functions.R
@@ -6,7 +6,7 @@
 #' Note that some data is not available for some countries (mostly non-OECD).
 #'
 #' @param countries A string of iso 3-letter country codes
-#' @param isocode_colname See `PFUDatabase::socioecon_cols`.
+#' @param isocode_colname See `PFUPipeline::socioecon_cols`.
 #'
 #' @return A data frame containing socioeconomic data from pwt10 for a set of countries.
 
@@ -16,7 +16,7 @@
 #' countries <- c("GBR")
 #' all_pwt_data <- get_all_pwt_data(countries = countries)
 get_all_pwt_data <- function(countries,
-                             isocode_colname = PFUDatabase::socioecon_cols$isocode_colname) {
+                             isocode_colname = PFUPipeline::socioecon_cols$isocode_colname) {
   
   # Get all pwt10 data and filter for countries in the string countries
   pwt10_data <- pwt10::pwt10.0 %>%
@@ -55,7 +55,7 @@ get_all_pwt_data <- function(countries,
 #'
 #' @param pwt10_data A data frame containing all pwt10 data for at least one country,
 #'                   usually obtained from `get_all_pwt_data()`.
-#' @param isocode_colname,year_colname,rgdpe_colname,rgdpo_colname,rgdpna_colname,emp_colname,avh_colname,hc_colname,rnna_colname,rkna_colname,K_colname,Kserv_colname,L_colname,Ladj_colname See `PFUDatabase::socioecon_cols`.
+#' @param isocode_colname,year_colname,rgdpe_colname,rgdpo_colname,rgdpna_colname,emp_colname,avh_colname,hc_colname,rnna_colname,rkna_colname,K_colname,Kserv_colname,L_colname,Ladj_colname See `PFUPipeline::socioecon_cols`.
 #' @param country_colname,Year_colname See `IEATools::iea_cols`.
 #'
 #' @return A data frame containing three GDP metrics, Labor, Adjusted Labor,
@@ -69,20 +69,20 @@ get_all_pwt_data <- function(countries,
 get_L_K_GDP_data <- function(pwt10_data,
                              country_colname = IEATools::iea_cols$country,
                              Year_colname= IEATools::iea_cols$year,
-                             year_colname= PFUDatabase::socioecon_cols$year_colname,
-                             isocode_colname = PFUDatabase::socioecon_cols$isocode_colname,
-                             rgdpe_colname = PFUDatabase::socioecon_cols$rgdpe_colname,
-                             rgdpo_colname = PFUDatabase::socioecon_cols$rgdpo_colname,
-                             rgdpna_colname = PFUDatabase::socioecon_cols$rgdpna_colname,
-                             emp_colname = PFUDatabase::socioecon_cols$emp_colname,
-                             avh_colname = PFUDatabase::socioecon_cols$avh_colname,
-                             hc_colname = PFUDatabase::socioecon_cols$hc_colname,
-                             rnna_colname = PFUDatabase::socioecon_cols$rnna_colname,
-                             rkna_colname = PFUDatabase::socioecon_cols$rkna_colname,
-                             K_colname = PFUDatabase::socioecon_cols$K_colname,
-                             Kserv_colname = PFUDatabase::socioecon_cols$Kserv_colname,
-                             L_colname = PFUDatabase::socioecon_cols$L_colname,
-                             Ladj_colname = PFUDatabase::socioecon_cols$Ladj_colname) {
+                             year_colname= PFUPipeline::socioecon_cols$year_colname,
+                             isocode_colname = PFUPipeline::socioecon_cols$isocode_colname,
+                             rgdpe_colname = PFUPipeline::socioecon_cols$rgdpe_colname,
+                             rgdpo_colname = PFUPipeline::socioecon_cols$rgdpo_colname,
+                             rgdpna_colname = PFUPipeline::socioecon_cols$rgdpna_colname,
+                             emp_colname = PFUPipeline::socioecon_cols$emp_colname,
+                             avh_colname = PFUPipeline::socioecon_cols$avh_colname,
+                             hc_colname = PFUPipeline::socioecon_cols$hc_colname,
+                             rnna_colname = PFUPipeline::socioecon_cols$rnna_colname,
+                             rkna_colname = PFUPipeline::socioecon_cols$rkna_colname,
+                             K_colname = PFUPipeline::socioecon_cols$K_colname,
+                             Kserv_colname = PFUPipeline::socioecon_cols$Kserv_colname,
+                             L_colname = PFUPipeline::socioecon_cols$L_colname,
+                             Ladj_colname = PFUPipeline::socioecon_cols$Ladj_colname) {
   
   # Select columns
   L_K_GDP_data <- pwt10_data %>%